requirements.txt
torch>=1.10
torchvision>=0.11
torchmetrics>=1.1.2
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=9.0.0
tqdm>=4.65.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
ptflops>=0.7.2
swanlab>=0.4.3
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==6.0.1
polars==1.33.1
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
ptflops==0.7.5
ptyprocess==0.7.0
pure_eval==0.2.3
pycm==4.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyecharts==2.0.9
Pygments==2.19.1
pykan==0.2.8
pynndescent==0.5.13
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytorch-lightning==2.5.1
pytz==2023.4
PyWavelets==1.6.0
PyYAML==5.4.1
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
RestrictedPython==8.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rotary-embedding-torch==0.8.6
rpds-py==0.24.0
s3transfer==0.11.4
safetensors==0.5.2
scanpy==1.10.3
scikit-image==0.24.0
scikit-learn==1.6.1
scikit-video==1.1.11
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.22.0
session_info==1.0.1
setproctitle==1.3.5
shap==0.47.1
shellingham==1.5.4
SimpleITK==2.4.1
simplejson==3.20.2
simsimd==6.2.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.5
stdlib-list==0.11.1
stringzilla==3.12.3
swanlab==0.7.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.8.0
torch-fidelity==0.3.0
torch-geometric==2.6.1
torchaudio==2.8.0
torchcam==0.4.0
torchdiffeq==0.2.5
torchinfo==1.8.0
torchio==0.20.4
torchmetrics==1.6.1
torchsummary==1.5.1
torchvision==0.23.0
torchviz==0.0.3
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.50.0
triton==3.4.0
ttach==0.0.3
typer==0.15.1
types-python-dateutil==2.9.0.20250708
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.204
ultralytics-thop==2.0.17
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
vit-pytorch==1.10.1
volumentations-3D==1.0.4
vtk==9.4.2
wandb==0.19.7
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yapf==0.43.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
